RFC conformity tweaks
Trace header checking is being moved to the server with individual endpoints having the option to override the server's return values.

HEAD now functions the same as GET but does not return a body.

Endpoints must now define their allowed methods via the getAllowedMethods array. The current plan is to allow the Endpoint class to serve as a proper base template that all future endpoints can choose to utilize with only the interface being required.